Move the sample configuration file to top directory
Move the sample configuration file to top directory to avoid
user can't  find it, move the config module to common subdirectory
as it is a common function.

Rename the configuration parameter from protocolconfiguration to
configuration as it used by all modules intead of protocol only
