+++ common/log/log.go
@@ -1,7 +1,7 @@
 package log
 
 import (
-	"dad-go/config"
+	"dad-go/common/config"
 	"bytes"
 	"errors"
 	"fmt"
