+++ net/net.go
@@ -2,7 +2,7 @@ package net
 
 import (
 	. "dad-go/common"
-	"dad-go/config"
+	"dad-go/common/config"
 	"dad-go/core/ledger"
 	"dad-go/core/transaction"
 	"dad-go/crypto"
